backend/requirements.txt
fastapi
uvicorn
python-dotenv
motor
openai
sqlalchemy
passlib
alembic
python-multipart
pydub
PyJWT
bcrypt